Cleaning up `performance_counter` package - heavy renaming and abstracting out functionality across `psutil` and peformance counters classes.
